+++ src/main/java/com/codiary/backend/domain/comment/service/CommentService.java
@@ -61,6 +61,38 @@ public class CommentService {
         return commentRepository.save(comment);
     }
 
+    public Comment replyToComment(Long commentId, Long replierId, CommentRequestDTO.CommentDTO request) {
+        // validation: 사용자 유무, 댓글 유무
+        Member replier = memberRepository.findById(replierId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
+        Post post = postRepository.findById(comment.getPost().getPostId())
+                .orElseThrow(() -> new GeneralException(ErrorStatus.POST_NOT_FOUND));
+
+        // validation: 사용자 권한 확인
+        if (post.getPostAccess().equals(PostAccess.MEMBER) && post.getMember() != replier) {
+            throw new GeneralException(ErrorStatus.COMMENT_CREATE_UNAUTHORIZED);
+        } else if (post.getPostAccess().equals(PostAccess.TEAM)) {
+            Team teamOfPost = teamRepository.findByIdWithTeamMemberList(post.getTeam().getTeamId())
+                    .orElseThrow(() -> new TeamHandler(ErrorStatus.TEAM_NOT_FOUND));
+            if (!teamRepository.isTeamMember(teamOfPost, replier)) {
+                throw new GeneralException((ErrorStatus.COMMENT_CREATE_UNAUTHORIZED));
+            }
+        }
+
+        // business logic: 대댓글 생성
+        Comment reply = Comment.builder()
+                .commentBody(request.commentBody())
+                .member(replier)
+                .parent(comment)
+                .post(null)
+                .build();
+
+        // response: 대댓글 반환
+        return commentRepository.save(reply);
+    }
+
     public String deleteComment(Long commentId, Long memberId) {
         // validation: 사용자, comment 유무 확인
         // + 사용자가 해당 댓글에 대한 권한 있는지
@@ -79,13 +111,12 @@ public class CommentService {
         return "성공적으로 삭제되었습니다!";
     }
 
-    @Transactional
     public Comment updateComment(Long commentId, Long memberId, CommentRequestDTO.CommentDTO request) {
         // validation: 사용자, comment 유무 확인
         // + 사용자가 해당 댓글에 대한 권한 있는지
         Member requester = memberRepository.findById(memberId)
                 .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-        Comment comment = commentRepository.findById(commentId)
+        Comment comment = commentRepository.findByIdWithReplies(commentId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
         if (comment.getMember() != requester) {
             throw new GeneralException(ErrorStatus.COMMENT_UPDATE_UNAUTHORIZED);
@@ -117,10 +148,38 @@ public class CommentService {
         }
 
         // business logic: 댓글 조회
-        List<Comment> comments = commentRepository.findByPostWithMemberInfoAndRepliesOrderByCreatedAtAsc(postId,
-                pageable);
+        List<Comment> comments
+                = commentRepository.findByPostWithMemberInfoAndRepliesOrderByCreatedAtAsc(postId, pageable);
 
         // response: comment list 반환
         return comments;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Comment> getReplies(Long commentId, Long requesterId, Pageable pageable) {
+        // validation: 사용자, 댓글 유무 확인
+        Member requester = memberRepository.findById(requesterId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
+        Post post = postRepository.findById(comment.getPost().getPostId())
+                .orElseThrow(() -> new GeneralException(ErrorStatus.POST_NOT_FOUND));
+
+        // validation: 사용자 권한 확인
+        if (post.getPostAccess().equals(PostAccess.MEMBER) && post.getMember() != requester) {
+            throw new GeneralException(ErrorStatus.COMMENT_CREATE_UNAUTHORIZED);
+        } else if (post.getPostAccess().equals(PostAccess.TEAM)) {
+            Team teamOfPost = teamRepository.findByIdWithTeamMemberList(post.getTeam().getTeamId())
+                    .orElseThrow(() -> new TeamHandler(ErrorStatus.TEAM_NOT_FOUND));
+            if (!teamRepository.isTeamMember(teamOfPost, requester)) {
+                throw new GeneralException((ErrorStatus.COMMENT_CREATE_UNAUTHORIZED));
+            }
+        }
+
+        // business logic: 대댓글 조회
+        List<Comment> replies = commentRepository.findByParentWithMemberInfoOrderByCreatedAtAsc(commentId, pageable);
+
+        // response
+        return replies;
     }
 }
